requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0
ftfy==6.1.3
dashscope==1.10.0
numpy==1.24.3
huggingface_hub==0.17.3
Pillow>=9.0.0
tqdm>=4.64.1
einops>=0.6.0
replicate>=0.11.0